% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STARcounts.R
\name{write.table_STARcounts}
\alias{write.table_STARcounts}
\title{STAR counts}
\usage{
write.table_STARcounts(
  results_path = "./results",
  table_name = "STARcountDFeByg.xls",
  dbg = 0
)
}
\arguments{
\item{results_path}{Path to the STAR Aligner count files}

\item{table_name}{Name of the output file including the .xls extension}

\item{dbg}{Optional debug level defaults to 0}
}
\value{

}
\description{
This function loads STAR Aligner count files *ReadsPerGene.out.tab$ from results_path, accumulates the
data into a data frame, then writes the results to table_name.
Some code from biostars 241602.
Some code from stackoverflow 13762224.
annotation.gtf must already exist before the mapping step, ./gffread -T annotation.gff -o annotation.gtf
_SampleName_.ReadsPerGene.out.tab must already exist (created during mapping step)
}
\examples{
write.table_STARcounts(results_path="./results", table_name="STARcountDFeByg.xls")
countDF <- read.delim("results/STARcountDFeByg.xls", row.names=1, check.names=FALSE)
head(countDF)
}
